server/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
pymupdf==1.21.1
pytesseract==0.3.10
Pillow==9.4.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
aiohttp==3.8.4
requests==2.28.2
Werkzeug==2.2.3
SQLAlchemy==2.0.4
celery==5.2.7
redis==4.5.1
